+++ react/components/UploadingScreen/index.tsx
@@ -1,0 +1,2 @@
+export { default as ValidationScreen } from './UploadingScreen'
+export type { UploadingScreenProps } from './UploadingScreen'