+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Refactor Bulk import Uploading Modal to Async Validation
+
 ## [1.29.2] - 2024-01-12
 
 ### Fixed